MISRA Compliance Update (#132)
* Changes to MISRA.md to reflect new format.
Added two variables to caseInsensitiveStringCmp function for MISRA compliance.
Changed in line comments to reflect new MISRA.md format.
Modified misra.config reasons.

* The cast to (unsigned char) before passing the char[index] value into
toupper() was causing the CBMC test to fail. Through much trial and
error it appears there isn't a way around it. So adding in a MISRA
exception for the rule related to that cast.
* Using DNDEBUG when building coverity_analysis target for removal of assert.

Co-authored-by: Soren Ptak <skptak@amazon.com>